# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Citibike demo API",
    "description" : "API for the Citibike Demo",
    "version" : "1.0.1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/Snowflake5/Citibike/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "tags" : [ {
    "name" : "developers"
  } ],
  "paths" : {
    "/stations" : {
      "get" : {
        "tags" : [ "developers" ],
        "summary" : "searches stations",
        "description" : "By passing in the appropriate options, you can search for\navailable stations in the system\n",
        "operationId" : "searchInventory",
        "parameters" : [ {
          "name" : "searchString",
          "in" : "query",
          "description" : "pass an optional search string for looking up station names",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StationItem"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "post" : {
        "tags" : [ "developers" ],
        "summary" : "creates a station",
        "description" : "Adds an item to the system",
        "operationId" : "addStation",
        "requestBody" : {
          "description" : "Station to add",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StationItem"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "station created"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "an existing item already exists"
          }
        }
      }
    },
    "/stations/{stationId}" : {
      "get" : {
        "tags" : [ "developers" ],
        "parameters" : [ {
          "name" : "stationId",
          "in" : "path",
          "description" : "ID of station",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StationItem"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "developers" ],
        "parameters" : [ {
          "name" : "stationId",
          "in" : "path",
          "description" : "ID of station",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StationItem"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "developers" ],
        "parameters" : [ {
          "name" : "stationId",
          "in" : "path",
          "description" : "ID of station",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StationItem"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "StationItem" : {
        "required" : [ "latitude", "longitude", "name" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "number",
            "example" : 450.0
          },
          "name" : {
            "type" : "string",
            "example" : "W 49 St & 8 Ave"
          },
          "latitude" : {
            "type" : "number",
            "example" : 40.761636025
          },
          "longitude" : {
            "type" : "number",
            "example" : -73.988441025
          },
          "comment" : {
            "type" : "string",
            "example" : "Bike station at W 49 St & 8 Ave"
          }
        },
        "example" : {
          "latitude" : 40.761636025,
          "name" : "W 49 St & 8 Ave",
          "comment" : "Bike station at W 49 St & 8 Ave",
          "id" : 450.0,
          "longitude" : -73.988441025
        }
      }
    }
  }
}